backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
sympy==1.13.3
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
python-dotenv
psycopg2
google-generativeai
httpx
